+++ src/command/get_all_databases.rs
@@ -1,7 +1,8 @@
 use hyper;
 
 use client::{self, ClientState};
-use error::{self, Error};
+use error::Error;
+use transport::{self, Command, Request};
 
 #[doc(hidden)]
 pub fn new_get_all_databases(client_state: &ClientState) -> GetAllDatabases {
@@ -26,25 +27,28 @@ impl<'a> GetAllDatabases<'a> {
     /// This command has no specific errors.
     ///
     pub fn run(self) -> Result<Vec<String>, Error> {
+        transport::run_command(self)
+    }
+}
 
-        let mut resp = {
-            use hyper::mime::{Mime, TopLevel, SubLevel};
-            let mut u = self.client_state.uri.clone();
-            u.path_mut().unwrap()[0] = "_all_dbs".to_string();
-            try!(
-                self.client_state.http_client.get(u)
-                .header(hyper::header::Accept(vec![
-                    hyper::header::qitem(
-                        Mime(TopLevel::Application, SubLevel::Json, vec![]))]))
-                .send()
-                .or_else(|e| {
-                    Err(Error::Transport {
-                        cause: error::TransportCause::Hyper(e),
-                    })
-                })
-            )
+impl<'a> Command for GetAllDatabases<'a> {
+
+    type Output = Vec<String>;
+
+    fn make_request(self) -> Result<Request, Error> {
+        let uri = {
+            let mut uri = self.client_state.uri.clone();
+            uri.path_mut().unwrap()[0] = "_all_dbs".to_string();
+            uri
         };
+        let req = try!(Request::new(hyper::Get, uri))
+            .accept_application_json();
+        Ok(req)
+    }
 
+    fn take_response(mut resp: hyper::client::Response)
+        -> Result<Self::Output, Error>
+    {
         match resp.status {
             hyper::status::StatusCode::Ok => {
                 let s = try!(client::read_json_response(&mut resp));